refactor(fleet-wiki): extract the cowork engine into a fleet-wiki subpackage
- Move the Cowork domain core out of the console host into
  @dotobokuri/fleet-wiki/cowork: session DTO/record types, the
  in-memory CoworkStore, the one-shot CoworkService, and the scoped
  MCP runtime now live beside the draft tools they orchestrate.
- Fleet Console keeps only its adapters: the HTTP/SSE route handler,
  origin gates, the options endpoint (core-unified-agent registry),
  the gateway service cache, and the browser UI. Console api-types
  re-exports the cowork DTOs type-only from the subpackage.
- fleet-wiki gains a ./cowork export (own tsup entry) and a
  core-unified-agent dependency; engine tests move to fleet-wiki
  (219 tests), console keeps adapter and client tests.
