+++ .changelog.d/pr-299.md
@@ -7,5 +7,5 @@
 
 ### fleet-core
 #### Added
-- [fleet-wiki] Add session-scoped Cowork draft tools for reading, exact-replace editing, and full rewriting of one injected draft, kept out of the global wiki tool registry.
-  ko: 주입된 초안 하나를 읽고, 정확 치환으로 편집하고, 전체 재작성하는 세션 한정 Cowork draft 도구를 추가하며 전역 위키 도구 레지스트리에는 등록하지 않습니다.
+- [fleet-wiki] Add a cowork subpackage housing the terminal-free AI draft-editing engine: an in-memory session store, a one-shot agent service, a scoped MCP runtime, and session-scoped draft tools kept out of the global wiki tool registry, consumed by Fleet Console through its HTTP adapter.
+  ko: 터미널 없는 AI 초안 편집 엔진을 담는 cowork 서브패키지를 추가합니다 — 인메모리 세션 store, 원샷 에이전트 서비스, 스코프드 MCP 런타임, 전역 위키 도구 레지스트리에 등록되지 않는 세션 한정 draft 도구로 구성되며 Fleet Console은 HTTP 어댑터로 이를 소비합니다.
