+++ packages/fleet-wiki/tsup.config.ts
@@ -2,7 +2,8 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: {
-    index: "src/index.ts"
+    index: "src/index.ts",
+    cowork: "src/cowork/index.ts"
   },
   format: ["esm", "cjs"],
   dts: true,
